check if tests are ran with correct version of glib

+++ .jenkins.sh
@@ -29,6 +29,10 @@ BALDE_BUILD_DIR="${BALDE_SRC_DIR}/build"
 export PKG_CONFIG_LIBDIR="${GLIB_BASE_DIR}/lib/pkgconfig"
 export PATH="${GLIB_BASE_DIR}/bin:${PATH}"
 
+GLIB_VERSION_PKGCONFIG="$(pkg-config --modversion glib-2.0)"
+if [[ "${GLIB_VERSION_PKGCONFIG}" != "${GLIB_VERSION}" ]]; then
+    exit 1
+fi
 
 rm -rf "${BALDE_BUILD_DIR}"
